[Flux] TPU Optimizations & Speedups for Flux.1-dev, Flux.1-schnell, and Flux.2-klein
- Implemented performance profiling, kernel fusion, instruction scheduling, and latency optimizations for the Flux model family (Flux.1-dev, Flux.1-schnell, Flux.2-klein-4B/9B) on Cloud TPU v6e (Trillium).
- Enabled ulysses_custom_fixed_m attention kernel with base-2 exponential scaling (use_base2_exp: True).
- Automatically routes heads_per_tile=1 for fixed-m attention dispatches (preserving heads_per_tile>1 tile batching for non-fixed-m kernels).
- Enabled LP LLO instruction scheduler in Pallas (use_experimental_scheduler: True).
- Fused Rotary Position Embeddings (RoPE) directly on (B, L, H, D) layout, eliminating 228 redundant transpose/swapaxes ops.
- Fused LayerNorm scale and shift calculations into single multiply-add ops.
- Maintained latents in native bfloat16 for VAE decoding, cutting VAE decode time by 54.2%.
- Fixed null check for image_rotary_emb in FlaxFluxAttention to prevent AttributeError.

+++ src/maxdiffusion/pipelines/flux/flux2klein_pipeline.py
@@ -30,7 +30,7 @@ from flax.linen import partitioning as nn_partitioning
 from maxdiffusion import max_logging
 from maxdiffusion.max_utils import device_put_replicated
 from ..pipeline_flax_utils import FlaxDiffusionPipeline
-from ...models.flux.transformers.transformer_flux_flax import Flux2KleinTransformer2DModel
+from ...models.flux.transformers.transformer_flux import Flux2KleinTransformer2DModel
 from ...models.vae_flax import FlaxAutoencoderKL
 from ...models.qwen3_flax import FlaxQwen3Model
 from ...schedulers.scheduling_flow_match_flax import FlaxFlowMatchScheduler, compute_empirical_mu
@@ -345,12 +345,12 @@ class FlaxFlux2KleinPipeline(FlaxDiffusionPipeline):
     # Apply Channel-wise Batch Normalization Scaling in packed sequence format (denormalize)
     vae_bn_mean_seq = vae_bn_mean.reshape(1, 1, 128)
     vae_bn_std_seq = vae_bn_std.reshape(1, 1, 128)
-    latents_bn = latents_jax * vae_bn_std_seq + vae_bn_mean_seq
+    latents_bn = (latents_jax * vae_bn_std_seq + vae_bn_mean_seq).astype(jnp.bfloat16)
 
-    # Unpack packed latents back to spatial grid
+    # Unpack packed latents back to spatial grid on TPU device
     latents_unpacked = unpack_latents(latents_bn, batch_size, 32, height, width)
 
-    # Decode VAE latents to RGB pixels
+    # Decode VAE latents to RGB pixels natively in BF16
     decoded_out = self._jitted_vae_decode(vae_params, latents_unpacked)
     # VAE output is in decoded_out.sample
     images_rgb = decoded_out.sample
